+++ drivers/md/dm-vdo/thread-utils.h
@@ -11,15 +11,10 @@
 #include <linux/jiffies.h>
 #include <linux/mutex.h>
 #include <linux/semaphore.h>
-#include <linux/wait.h>
 
 #include "errors.h"
 
 /* Thread and synchronization utilities for UDS */
-
-struct cond_var {
-	wait_queue_head_t wait_queue;
-};
 
 struct thread;
 
@@ -31,29 +26,7 @@ void uds_perform_once(atomic_t *once_state, void (*function) (void));
 
 int uds_join_threads(struct thread *thread);
 
-static inline int __must_check uds_init_cond(struct cond_var *cv)
-{
-	init_waitqueue_head(&cv->wait_queue);
-	return UDS_SUCCESS;
-}
-
-static inline void uds_signal_cond(struct cond_var *cv)
-{
-	wake_up(&cv->wait_queue);
-}
-
-static inline void uds_broadcast_cond(struct cond_var *cv)
-{
-	wake_up_all(&cv->wait_queue);
-}
-
-void uds_wait_cond(struct cond_var *cv, struct mutex *mutex);
-
 /* FIXME: all below wrappers should be removed! */
-
-static inline void uds_destroy_cond(struct cond_var *cv)
-{
-}
 
 static inline int __must_check uds_init_mutex(struct mutex *mutex)
 {
@@ -75,6 +48,5 @@ static inline void uds_unlock_mutex(struct mutex *mutex)
 {
 	mutex_unlock(mutex);
 }
-
 
 #endif /* UDS_THREADS_H */
